+++ packages/core/src/Drivers/SystemTaxDriver.php
@@ -6,8 +6,8 @@ use Lunar\Actions\Taxes\GetTaxZone;
 use Lunar\Base\Addressable;
 use Lunar\Base\Purchasable;
 use Lunar\Base\TaxDriver;
-use Lunar\Base\ValueObjects\TaxBreakdown;
-use Lunar\Base\ValueObjects\TaxBreakdownAmount;
+use Lunar\Base\ValueObjects\Cart\TaxBreakdown;
+use Lunar\Base\ValueObjects\Cart\TaxBreakdownAmount;
 use Lunar\DataTypes\Price;
 use Lunar\Models\CartLine;
 use Lunar\Models\Currency;
