+++ packages/core/tests/Stubs/TestTaxDriver.php
@@ -5,8 +5,8 @@ namespace Lunar\Tests\Stubs;
 use Lunar\Base\Addressable;
 use Lunar\Base\Purchasable;
 use Lunar\Base\TaxDriver;
-use Lunar\Base\ValueObjects\TaxBreakdown;
-use Lunar\Base\ValueObjects\TaxBreakdownAmount;
+use Lunar\Base\ValueObjects\Cart\TaxBreakdown;
+use Lunar\Base\ValueObjects\Cart\TaxBreakdownAmount;
 use Lunar\DataTypes\Price;
 use Lunar\Models\CartLine;
 use Lunar\Models\Currency;
