+++ packages/core/tests/Unit/Facades/TaxesTest.php
@@ -4,7 +4,7 @@ namespace Lunar\Tests\Unit\Facades;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Lunar\Base\TaxManagerInterface;
-use Lunar\Base\ValueObjects\TaxBreakdown;
+use Lunar\Base\ValueObjects\Cart\TaxBreakdown;
 use Lunar\Facades\Taxes;
 use Lunar\Tests\Stubs\TestTaxDriver;
 use Lunar\Tests\TestCase;
